+++ src/test/java/lotto/domain/LottoNumberTest.java
@@ -9,7 +9,7 @@ import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException
 
 public class LottoNumberTest {
     
-    @ParameterizedTest(name = "1~45의 숫자를 입력할 경우 정상적으로 로또 번호 생성")
+    @ParameterizedTest(name = "범위 안의 숫자를 입력할 경우 정상적으로 로또 번호 생성")
     @ValueSource(ints = {1, 2, 3, 43, 44, 45})
     public void initTest(int lottoNum) {
         
@@ -20,7 +20,7 @@ public class LottoNumberTest {
         assertThatCode(throwingCallable).doesNotThrowAnyException();
     }
     
-    @ParameterizedTest(name = "1~45 외의 숫자를 입력할 경우 예외 발생")
+    @ParameterizedTest(name = "범위 외의 숫자를 입력할 경우 예외 발생")
     @ValueSource(ints = {-2, -1, 0, 46, 47, 48})
     public void init_OutOfBoundsNumber_ExceptionThrown(int lottoNum) {
